fix: 루틴이 없는 경우 수행하기를 클릭했을 때 문제 해결
